Execute Ruby code inside config files
It is pretty helpful to have values in config files dependent on instance
variables. Having so, we can store all config in the environment, which
is a nice 12-factor app practice. Therefore, the configuration files
become a kind of anticoruption layer with a tiny logic of setting
default values when the env variable isn't set.

This commit unlocks such functionality.